Disregard email, just draw art
